feature(add: protection routes get getById update delete create)

+++ build/utils/graphqlTypes/fields/delete.js
@@ -25,8 +25,11 @@ exports.default = {
         },
     }),
     description: 'Filtered location list',
-    resolve(_, args) {
+    resolve(_, args, context) {
         return __awaiter(this, void 0, void 0, function* () {
+            const auth = context();
+            if (!auth.user)
+                throw new Error('UNAHUTORIZED');
             const deletedData = yield service.delete(args.id, args.controller);
             return deletedData;
         });
